requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.12
Jinja2==3.1.4
python-dotenv==1.0.1
fitparse==1.2.0
google-generativeai==0.7.2
